refactor(adapters): address review
unflatten exports
