+++ python/beeai_framework/adapters/__init__.py
@@ -12,29 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from beeai_framework.adapters.amazon_bedrock.backend.chat import AmazonBedrockChatModel
-from beeai_framework.adapters.anthropic.backend.chat import AnthropicChatModel
-from beeai_framework.adapters.azure_openai.backend.chat import AzureOpenAIChatModel
-from beeai_framework.adapters.groq.backend.chat import GroqChatModel
-from beeai_framework.adapters.langchain.tools import LangChainTool, LangChainToolRunOptions
-from beeai_framework.adapters.litellm.chat import LiteLLMChatModel
-from beeai_framework.adapters.ollama.backend.chat import OllamaChatModel
-from beeai_framework.adapters.openai.backend.chat import OpenAIChatModel
-from beeai_framework.adapters.vertexai.backend.chat import VertexAIChatModel
-from beeai_framework.adapters.watsonx.backend.chat import WatsonxChatModel
-from beeai_framework.adapters.xai.backend.chat import XAIChatModel
-
-__all__ = [
-    "AmazonBedrockChatModel",
-    "AnthropicChatModel",
-    "AzureOpenAIChatModel",
-    "GroqChatModel",
-    "LangChainTool",
-    "LangChainToolRunOptions",
-    "LiteLLMChatModel",
-    "OllamaChatModel",
-    "OpenAIChatModel",
-    "VertexAIChatModel",
-    "WatsonxChatModel",
-    "XAIChatModel",
-]
