+++ python/beeai_framework/adapters/ollama/__init__.py
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from beeai_framework.adapters.ollama.backend.chat import OllamaChatModel
 
+__all__ = ["OllamaChatModel"]
