+++ python/docs/backend.md
@@ -73,7 +73,7 @@ import traceback
 
 from pydantic import BaseModel, Field
 
-from beeai_framework.adapters import WatsonxChatModel
+from beeai_framework.adapters.watsonx import WatsonxChatModel
 from beeai_framework.backend import ChatModel, MessageToolResultContent, ToolMessage, UserMessage
 from beeai_framework.cancellation import AbortSignal
 from beeai_framework.errors import AbortError, FrameworkError
@@ -259,7 +259,7 @@ import asyncio
 import sys
 import traceback
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.backend import UserMessage
 from beeai_framework.errors import FrameworkError
 from examples.helpers.io import ConsoleReader
