+++ python/docs/memory.md
@@ -119,7 +119,7 @@ import asyncio
 import sys
 import traceback
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.backend import AssistantMessage, SystemMessage, UserMessage
 from beeai_framework.errors import FrameworkError
 from beeai_framework.memory import UnconstrainedMemory
@@ -360,7 +360,7 @@ import math
 import sys
 import traceback
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.backend import Role, SystemMessage, UserMessage
 from beeai_framework.errors import FrameworkError
 from beeai_framework.memory import TokenMemory
