+++ python/examples/agents/experimental/human.py
@@ -2,7 +2,7 @@ import asyncio
 import sys
 import traceback
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.agents import AgentExecutionConfig
 from beeai_framework.agents.react import ReActAgent
 from beeai_framework.errors import FrameworkError
