+++ python/examples/backend/chat.py
@@ -2,7 +2,7 @@ import asyncio
 import sys
 import traceback
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.backend import UserMessage
 from beeai_framework.errors import FrameworkError
 from examples.helpers.io import ConsoleReader
