+++ python/examples/backend/providers/watsonx.py
@@ -5,7 +5,7 @@ import traceback
 
 from pydantic import BaseModel, Field
 
-from beeai_framework.adapters import WatsonxChatModel
+from beeai_framework.adapters.watsonx import WatsonxChatModel
 from beeai_framework.backend import ChatModel, MessageToolResultContent, ToolMessage, UserMessage
 from beeai_framework.cancellation import AbortSignal
 from beeai_framework.errors import AbortError, FrameworkError
