+++ python/examples/tools/python_tool.py
@@ -6,7 +6,7 @@ import traceback
 
 from dotenv import load_dotenv
 
-from beeai_framework.adapters import OllamaChatModel
+from beeai_framework.adapters.ollama import OllamaChatModel
 from beeai_framework.agents.react import ReActAgent
 from beeai_framework.errors import FrameworkError
 from beeai_framework.memory import UnconstrainedMemory
